// sanity.types.ts
export type Order = {
  status: string;
  _id: string;
  _type: "order";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  userId?: string;
  firstName?: string;
  lastName?: string;
  country?: string;
  address?: string;
  city?: string;
  state?: string;
  postcode?: string;
  phone?: string;
  email?: string;
  products?: Array<{
    slug: any;
    name?: string;
    price?: number;
    quantity?: number;
    _key: string;
  }>;
  subtotal?: number;
  total?: number;
  createdAt?: string;
};